Correct the handling of the weight matrix Wq.
The weight is taken into account by putting a final
stage a the end of the stack of problems, such as:
min ||Wq.ydot||

Also, a second signal is used to define this matrix.
Wq_diag that defines the diagonal element of the matrix.
If the Wq matrix is already provided, the vector Wq_diag
is not taken into account.